[SPARK-37823][INFRA][TESTS] Add `is-changed.py` dev script
